pyproject.toml
[tool.poetry]
name = "shattered"
version = "0.1.0"
description = "STOMP meets bottle.py"
authors = ["Jimmy Bradshaw <james.g.bradshaw@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
"stomp.py" = "~5.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
